Changing the package name and refactoring the package so that the node sends OS and Process messages (not complete)
The network_read and network_write on the process is not being filled.
The temperature is the same.
Nothing related to the storage has being implemented.

The psutil doesn't give any information about the network usage from a process.
On ubuntu 16 the psutil version is 3.x, and the sensors_temperature() is not implemented.
Both will need something different from psutil to work.